wk5_6.001_ps_brainstorm
There is a problem set in 6.001 via edX to source code audit a password-guessing game written by the instructor to implement a function to randomly select a word from a list, if the letter is guessed correctly as contained in the word print the word's correctly guessed letter, print it otherwise print an underscore. I decided to re-write the whole game myself. I will be sitting 6.001 beginning June 2nd 2022, but was given access to the archived course with expired problem sets, exams and finger typing exercises. Very excited to take 6.001 and 6.002